refactor(mesh-v2): extract subscription-manager mixin
issue #566 の責務分割 4/N。WebSocket subscription (onMessageInGroup)
受信ロジックとメッセージのディスパッチを `subscription-manager.js` に
切り出し。

切り出し対象:
- startSubscriptions
- stopSubscriptions
- handleDataUpdate (subscription / pollGroupData / fetchAllNodesData の
  3 経路から呼ばれる共通取り込み口)
- handleBatchEvent

mesh-service.js: 1182 → 1093 行 (-89)
subscription-manager.js: 100 行 (新規)
